requirements.txt
fastapi
uvicorn
python-dotenv
google-cloud-dialogflow